// lancewells_github_io/src/GamePage/Interfaces/IGameRoom.ts
import { TRoomDecor } from '../Types/TRoomDecor';
import { TCharacterDisplay } from '../Types/TCharacterDisplay';

export interface IGameRoom {
    /**
     * The ID for the room. This won't be accessed much but is vital when telling any room creation
     * callers what we just made.
     */
    readonly RoomId: string;

    /**
     * The name of the room. Will be displayed in the room area.
     */
    readonly RoomName: string;

    /**
     * Describes any current decorations in the game room display.
     */
    GameRoomDecor: TRoomDecor;

    /**
     * A list of characters and their relevant visible media.
     */
    readonly Characters: TCharacterDisplay[];
}